refactor: aws service wrapper
