+++ flock/src/aws/mod.rs
@@ -1,0 +1,23 @@
+// Copyright (c) 2020-present, UMD Database Group.
+//
+// This program is free software: you can use, redistribute, and/or modify
+// it under the terms of the GNU Affero General Public License, version 3
+// or later ("AGPL"), as published by the Free Software Foundation.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+// FITNESS FOR A PARTICULAR PURPOSE.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+//! This module contains all wrapped functions of the AWS services such as
+//! Lambda, DynamoDB, S3, etc. Flock uses the AWS services to build the
+//! distributed query engine.
+
+pub mod cloudwatch;
+pub mod dynamodb;
+pub mod efs;
+pub mod lambda;
+pub mod s3;
+pub mod sqs;
