+++ flock/src/runtime/context.rs
@@ -14,6 +14,7 @@
 //! When the lambda function is called for the first time, it deserializes the
 //! corresponding execution context from the cloud environment variable.
 
+use crate::aws::s3;
 use crate::datasink::DataSinkType;
 use crate::encoding::Encoding;
 use crate::error::{FlockError, Result};
@@ -23,13 +24,9 @@ use datafusion::physical_plan::empty::EmptyExec;
 use datafusion::physical_plan::memory::MemoryExec;
 use datafusion::physical_plan::ExecutionPlan;
 use datafusion::physical_plan::{collect, collect_partitioned};
-use rusoto_core::Region;
-use rusoto_s3::GetObjectRequest;
-use rusoto_s3::{S3Client, S3};
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
 use std::collections::VecDeque;
-use std::io::Read;
 use std::sync::Arc;
 
 type CloudFunctionName = String;
@@ -152,27 +149,7 @@ impl ExecutionContext {
                 if self.plan_s3_idx.is_some() {
                     println!("Loading plan from S3 {:?}", self.plan_s3_idx);
                     let (bucket, key) = self.plan_s3_idx.as_ref().unwrap();
-                    let s3 = S3Client::new(Region::default());
-                    let body = s3
-                        .get_object(GetObjectRequest {
-                            bucket: bucket.clone(),
-                            key: key.clone(),
-                            ..Default::default()
-                        })
-                        .await
-                        .map_err(|e| FlockError::Internal(e.to_string()))?
-                        .body
-                        .take()
-                        .expect("body is empty");
-
-                    self.plan = tokio::task::spawn_blocking(move || {
-                        let mut buf = Vec::new();
-                        body.into_blocking_read().read_to_end(&mut buf).unwrap();
-                        serde_json::from_slice(&buf).unwrap()
-                    })
-                    .await
-                    .expect("failed to load plan from S3");
-
+                    self.plan = serde_json::from_slice(&s3::get_object(&bucket, &key).await?)?;
                     Ok(&mut self.plan)
                 } else {
                     panic!("The query plan is not stored in the environment variable and S3.");
